Add files via upload
Most of the functions so far work fine, however not all are well documented. The functions othello-place-token and othello-print-board are as yet unwritten. The function othello-capture (not to be confused with othello-capture-legal) is unfinished. Later in the development cycle I'll see about combining them into a single function, but not before the first demo. For clarity, othello-capture-legal takes in a board, and the row and column of a prospective move, and returns T or nil based on whether the move is legal, while othello-capture instead returns the actual board that move creates.